% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_get_data.R
\name{access_restricted_table}
\alias{access_restricted_table}
\title{access_restricted_table}
\usage{
access_restricted_table(entity_id, con)
}
\arguments{
\item{entity_id}{sidora table name of restricted tables (e.g. 'worker' etc.)}

\item{con}{database connection}
}
\description{
Some tables are restricted, e.g. the pandora_read user does not have access 
to certain columns. This function will use custom SQL queries to get all
useful (non-restricted) columns.
}
