% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_table_formatting.R
\name{as_update_existing}
\alias{as_update_existing}
\title{as_update_existing}
\usage{
as_update_existing(sidora_table)
}
\arguments{
\item{sidora_table}{a sidora table}
}
\description{
Converts a sidora table to a format that can be use with pandora table upload
interface
}
