% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{auxtablelookup}
\alias{auxtablelookup}
\title{auxtablelookup}
\description{
A named vector that includes all auxilary tables (e.g. Batch, Protocol, 
Location) and the corresponding column names that may be present in main 
tables
}
