% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_helper_functions.R
\name{str_to_colname}
\alias{str_to_colname}
\title{col_conversion}
\usage{
str_to_colname(entity_type, col)
}
\arguments{
\item{entity_type}{the table name in sidora format the requested column is derived from}

\item{col}{the column name to be selected}
}
\description{
Simple helper function to get around NSE issues with R.
 
Converts a table/column selection into a usable Sidora column object for
selecting columns by column name in a tibble.
}
