% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_lookup_in_pandora.R
\name{convert_default_lookup}
\alias{convert_default_lookup}
\title{convert_default_lookup}
\usage{
convert_default_lookup(df)
}
\arguments{
\item{df}{A Sidora/Pandora-Dataframe with standard column names.}
}
\value{
The converted dataframe with lookup-columns replaced by the actual values.
}
\description{
A convenience function which simply transforms a given Pandora-Dataframe using all
defined default lookups. Typically will convert a pandora back-end numeric ID to a 'human readable string' actually displayed on the pandora webpage.
}
