% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_filter_data.R
\name{filter_date}
\alias{filter_date}
\title{Filter a merged PANDORA table by date}
\usage{
filter_date(x, col, start = NA, end = NA, format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{data.frame}

\item{col}{character. String column in x (e.g. "Creation_Date.Site" or "Creation_Date.Individual") on which
the filter process should be applied}

\item{start}{Date. Start date of selection. Can also be supplied as a character (with format as defined in the format argument)}

\item{end}{Date. End date of selection. Can also be supplied as a character (with format as defined in the format argument)}

\item{format}{character. Date format definition relevant if start and end are supplied as characters}
}
\value{
filtered data.frame
}
\description{
Filter a merged PANDORA table by date
}
