% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_filter_data.R
\name{filter_pr_tag}
\alias{filter_pr_tag}
\title{Filter a merged PANDORA table by project or tag}
\usage{
filter_pr_tag(x, col, ins = c(), outs = c())
}
\arguments{
\item{x}{data.frame}

\item{col}{character. String column in x (e.g. "Projects.Individual" or "Tags.Individual") on which
the filter process should be applied}

\item{ins}{character vector. Projects or tags you want to see}

\item{outs}{character vector. Projects of tags you do not want to see}
}
\value{
filtered data.frame
}
\description{
Filter a merged PANDORA table by project or tag
}
