% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_table_formatting.R
\name{fix_logical_update_existing}
\alias{fix_logical_update_existing}
\title{fix_logical_update_existing}
\usage{
fix_logical_update_existing(sidora_table)
}
\arguments{
\item{x}{column name to wrap with asterisks}
}
\description{
fix_logical_update_existing
}
