% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_table_formatting.R
\name{format_as_update_existing}
\alias{format_as_update_existing}
\title{as_update_existing}
\usage{
format_as_update_existing(sidora_table)
}
\arguments{
\item{sidora_table}{a sidora table}
}
\description{
Converts a sidora table to a format that can be use with pandora table upload
interface
}
\examples{
\dontrun{
sidora_table <- get_df("TAB_Site", con)
format_as_update_existing(sidora_table,)
}

}
