% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_table_formatting.R
\name{format_table}
\alias{format_table}
\title{format_table}
\usage{
format_table(sidora_table, format)
}
\arguments{
\item{format}{a type of format to convert to.}

\item{table}{a sidora table}
}
\description{
This function takes a Sidora column-named table, and will re-format in 
various ways for saving in different contexts.
}
\examples{
\dontrun{
sidora_table <- get_df("TAB_Site", con)
format_table(sidora_table, "pandora_update_existing")
}

}
