% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_con}
\alias{get_con}
\title{Establish connection to database table}
\usage{
get_con(tab = sidora.core::pandora_tables, con)
}
\arguments{
\item{tab}{Name of table}

\item{con}{Database connection}
}
\value{
A table connection
}
\description{
Establish connection to database table
}
