% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_con_list}
\alias{get_con_list}
\title{Establish connections to multiple database tables}
\usage{
get_con_list(tab = sidora.core::pandora_tables, con)
}
\arguments{
\item{tab}{Names of tables}

\item{con}{Database connection}
}
\value{
A list of table connections
}
\description{
Establish connections to multiple database tables
}
