% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_get_data.R
\name{get_data}
\alias{get_data}
\alias{get_con}
\alias{get_con_list}
\alias{get_df}
\alias{get_df_list}
\title{Access and download tables from Pandora}
\usage{
get_con(tab = sidora.core::pandora_tables, con)

get_con_list(tab = sidora.core::pandora_tables, con)

get_df(
  tab = sidora.core::pandora_tables,
  con,
  cache = T,
  cache_dir = tempdir(),
  cache_max_age = 24 * 60 * 60
)

get_df_list(
  tab = sidora.core::pandora_tables,
  con,
  cache = T,
  cache_dir = tempdir(),
  cache_max_age = 24 * 60 * 60
)
}
\arguments{
\item{tab}{character vector. Names of tables}

\item{con}{database connection}

\item{cache}{logical. Should data be cached?}

\item{cache_dir}{character. Path to cache directory}

\item{cache_max_age}{numeric. Maximum age of cache in seconds}
}
\value{
connection, list of connections, data.frame or list of data.frames
}
\description{
See readme for more information.
}
