% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_df}
\alias{get_df}
\title{Download table from database}
\usage{
get_df(
  tab = c("TAB_Site", "TAB_Individual", "TAB_Sample", "TAB_Extract", "TAB_Library",
    "TAB_Capture", "TAB_Sequencing", "TAB_Raw_Data", "TAB_Sequencing_Sequencer",
    "TAB_Tag", "TAB_Project"),
  con,
  cache = T,
  cache_dir = tempdir(),
  cache_max_age = 24 * 60 * 60
)
}
\arguments{
\item{tab}{Name of table}

\item{con}{Database connection}

\item{cache}{Should data be cached?}

\item{cache_dir}{Path to cache directory}

\item{cache_max_age}{Maximum age of cache in seconds}
}
\value{
A dataframe
}
\description{
Download table from database
}
