% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_field_help}
\alias{get_field_help}
\title{Access field Pandora help tooltip information}
\usage{
get_field_help(tab, field, con)
}
\arguments{
\item{tab}{character. Name of the table of the field of interest in
"TAB_<table>" format. E.g. "TAB_Sample"}

\item{field}{character vector. Name of the field of interest. In Pandora
format E.g. "Sampled_Quantity"}

\item{con}{database connection object}
}
\value{
help text of the given field
}
\description{
This functions allows you to call the 'tooltip' help/documentation message
present on the Pandora webpage for a given entry.
}
