% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_helper_functions.R
\name{get_name_from_id}
\alias{get_name_from_id}
\title{get_name_from_id}
\usage{
get_name_from_id(con, query_tab, query_col, query_id, cache_dir)
}
\arguments{
\item{con}{a pandora connection}

\item{query_tab}{a sidora table name (e.g. 'site', 'individual' etc.)}

\item{query_col}{the column that that a pandora numeric Id to be converted is in}

\item{query_id}{a pandora numeric Id to be converted to the human readable 'string' version}

\item{cache_dir}{a cache directory}
}
\description{
Given a requested table, 'Id' column and a 'Id' integet will get the 
requested corresponding 'human readable' string version of the Id.
}
\details{
For example, given the.Batch ID 37 from the 'extract' sidora table,, would 
result in Ex06_KE_2015-11-19
}
\examples{

get_name_from_id(con = con, query_tab = "extract", query_col = "extract.Batch", query_id = 37)

}
