% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_helper_functions.R
\name{get_namecol_from_entity}
\alias{get_namecol_from_entity}
\title{get_namecol_from_entity}
\usage{
get_namecol_from_entity(entity_type)
}
\arguments{
\item{entity_type}{a valid entity type (e.g. site, sample, individual etc.)}
}
\description{
This gets the column of a given table, that contains the human-readable 'name' 
columns that will correspond to a numeric PANDORA ID.
}
