% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_pandora_connection.R
\name{get_pandora_connection}
\alias{get_pandora_connection}
\title{Get the connection object for the Pandora DB}
\usage{
get_pandora_connection(cred_file = ".credentials")
}
\arguments{
\item{cred_file}{character. Path to a credentials file containing four lines
listing the database host, the port of the database server, user and password,
respectively}
}
\value{
A Pandora DB connection object
}
\description{
Get the connection object for the Pandora DB
}
