% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_helper_functions.R
\name{id2string}
\alias{id2string}
\title{id2string}
\usage{
id2string(con, entity_type, id, cache_dir)
}
\arguments{
\item{entity_type}{a sidora table name (e.g. 'site', 'individual' etc.)}

\item{id}{the ID in the current table for which the 'string' version is requested}

\item{cache_dir}{}
}
\description{
Given a requested table and a 'Id' name, will get the requested corresponding 'Full_*_ID' string.
}
