% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{id_2_name_map}
\alias{id_2_name_map}
\title{id_2_name_map}
\description{
An named vector that includes all Pandora tables, but with the corresponding
Name column, that can be used to convert a 'Id' in a main table to the 
human-readable Name (as a string) column.
}
