% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_helper_functions.R
\name{is_sidoracol_auxid}
\alias{is_sidoracol_auxid}
\title{is_sidoracol_auxid}
\usage{
is_sidoracol_auxid(entity_type, col_name)
}
\arguments{
\item{entity_type}{a given sidora table of the column to be searched}

\item{col_name}{a sidora table column to be checked whether it is an numeric 'Id' column can be looked up in auxilary tables}
}
\description{
Provides logical value whether a given sidora column name is a valid
'Id' column that can be used for look-up in auxiliary tables
}
