% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{join_df_list}
\alias{join_df_list}
\title{Join multiple tables}
\usage{
join_df_list(df_list)
}
\arguments{
\item{df_list}{Named list of tables as produced by \code{get_df_list}}
}
\value{
Joined dataframe
}
\description{
Join multiple tables
}
