% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_data.R
\name{join_pandora_tables}
\alias{join_pandora_tables}
\title{Join multiple tables}
\usage{
join_pandora_tables(x)
}
\arguments{
\item{x}{Named list of tables or connections as produced by \code{get_df_list} or \code{get_con_list}}
}
\value{
Joined dataframe
}
\description{
Join multiple tables
}
