% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_table_formatting.R
\name{make_column_mandatory}
\alias{make_column_mandatory}
\title{make_column_mandatory}
\usage{
make_column_mandatory(x)
}
\arguments{
\item{x}{column name to wrap with asterisks}
}
\description{
make_column_mandatory
}
