% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_join_data.R
\name{make_complete_table_list}
\alias{make_complete_table_list}
\title{make_complete_table_list}
\usage{
make_complete_table_list(tabs, join_order_vector = sidora.core::pandora_tables)
}
\arguments{
\item{tabs}{test}

\item{join_order_vector}{test}
}
\description{
make_complete_table_list
}
