% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprod_make_progress_table.R
\name{make_progress_table}
\alias{make_progress_table}
\title{Make an individual-based progress table}
\usage{
make_progress_table(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame progress table
}
\description{
Make an individual-based progress table
}
