% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_helper_functions.R
\name{name_conversion}
\alias{name_conversion}
\alias{entity2table}
\alias{table2entity}
\alias{get_namecol_from_entity}
\title{Name conversion functions and lookup helpers}
\usage{
entity2table(entity_type)

table2entity(table_name)

get_namecol_from_entity(entity_type)
}
\arguments{
\item{entity_type}{character. A valid entity type (e.g. "site", "sample", "individual" etc.)}

\item{table_name}{character. A valid table name (e.g. "TAB_Site", "TAB_Sample" etc.)}
}
\description{
Convert Pandora entity type (e.g. "site", "sample", "individual" etc.) to 
Pandora table names (e.g. "TAB_Site", "TAB_Sample" etc.) and vice versa with
\code{entity2table} and \code{table2entity}. \cr\cr
Also look up id/name columns of entity types with \code{get_namecol_from_entity}.
}
