% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_lookup_in_pandora.R
\name{namecol_value_from_id}
\alias{namecol_value_from_id}
\title{get_name_from_id}
\usage{
namecol_value_from_id(sidora_col_name, query_id, con, cache_dir)
}
\arguments{
\item{sidora_col_name}{character. A sidora table column name}

\item{query_id}{a vector of pandora numeric Id(s) to be converted to the human readable 'string' version}

\item{con}{a pandora connection}

\item{cache_dir}{a cache directory}
}
\description{
Given a requested table, 'Id' column and a 'Id' integer will get the 
requested corresponding 'human readable' string version of the Id.
}
\details{
For example, given the.Batch ID 37 from the 'extract' sidora table, would 
result in Ex06_KE_2015-11-19
}
\examples{

namecol_value_from_id(sidora_col_name = "extract.Batch", query_id = 38, con = con)

}
