% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_table_formatting.R
\name{sidora_col_name_to_update_col_name}
\alias{sidora_col_name_to_update_col_name}
\title{sidora_col_name_to_update_col_name}
\usage{
sidora_col_name_to_update_col_name(x)
}
\arguments{
\item{x}{sidora column name to be converted to Pandora upload existing data column}
}
\description{
Reverts sidora-added underscores to spaces ni column names for pandora 
updating
}
