% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_helper_functions.R
\name{col_conversion}
\alias{col_conversion}
\alias{sidora_col_name_to_col_name}
\alias{col_name_to_sidora_col_name}
\alias{sidora_col_name_to_name}
\alias{col_name_to_name}
\title{col_conversion}
\usage{
sidora_col_name_to_col_name(sidora_col_name)

col_name_to_sidora_col_name(entity_type, col_name)

sidora_col_name_to_name(sidora_col_name)

col_name_to_name(entity_type, col_name)
}
\arguments{
\item{sidora_col_name}{character. A sidora table column name}

\item{entity_type}{character. An entity type (e.g. site, sample, individual etc.)}

\item{col_name}{character. A pandora table column name}
}
\description{
Simple helper functions to transform colnames and get around NSE issues with R.
}
\details{
str_to_colname converts a table/column selection into a usable Sidora column object for
selecting columns by column name in a tibble.
}
