% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_enforce_types.R
\name{enforce_types}
\alias{enforce_types}
\title{Enforce column types in a \strong{PANDORA table}}
\usage{
enforce_types(x, suppress_na_introduced_warnings = TRUE)
}
\arguments{
\item{x}{data.frame (PANDORA table)}

\item{suppress_na_introduced_warnings}{logical. Suppress warnings caused by data removal in
type transformation due to unrecognised columns}
}
\value{
data.frame (PANDORA table) with correct column types
}
\description{
See readme for more information.
}
