% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enforce_types.R
\name{enforce_types}
\alias{enforce_types}
\title{Enforce column types in a \strong{PANDORA table}}
\usage{
enforce_types(x, suppress_na_introduced_warnings = TRUE)
}
\arguments{
\item{x}{an object a PANDORA table}

\item{suppress_na_introduced_warnings}{suppress warnings caused by data removal in
type transformation due to unrecognised columns}
}
\value{
a PANDORA table
}
\description{
Enforce variable types in a \strong{PANDORA table}
}
\examples{
# initial situation
ex <- get_df("TAB_Site", con)

# modify variable/column type

# fix type with enforce_types()

}
