% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_table_formatting.R
\name{fix_logical_update_existing}
\alias{fix_logical_update_existing}
\title{fix_logical_update_existing}
\usage{
fix_logical_update_existing(sidora_table)
}
\arguments{
\item{sidora_table}{column name to wrap with asterisks}
}
\description{
Sidora converts on data download 'Yes/No' columns to logical column types 
for faster R processing. This converts these back, and also replaces 
non-Sample level ethically sensitive columns back to the dummy '0' data.
}
