% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_con}
\alias{get_con}
\title{Establish connection to database table}
\usage{
get_con(
  tab = c("TAB_Site", "TAB_Individual", "TAB_Sample", "TAB_Extract", "TAB_Library",
    "TAB_Capture", "TAB_Sequencing", "TAB_Raw_Data", "TAB_Sequencing_Sequencer",
    "TAB_Tag", "TAB_Project"),
  con
)
}
\arguments{
\item{tab}{Name of table that should be downloaded}

\item{con}{Database connection}
}
\value{
A table connection
}
\description{
Establish connection to database table
}
