% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_con_list}
\alias{get_con_list}
\title{Establish connections to multiple database tables}
\usage{
get_con_list(
  tab = c("TAB_Site", "TAB_Individual", "TAB_Sample", "TAB_Extract", "TAB_Library",
    "TAB_Capture", "TAB_Sequencing", "TAB_Raw_Data", "TAB_Sequencing_Sequencer",
    "TAB_Tag", "TAB_Project"),
  con,
  cache = T,
  cache_dir = tempdir(),
  cache_max_age = 24 * 60 * 60
)
}
\arguments{
\item{tab}{Names of tables}

\item{con}{Database connection}
}
\value{
A list of table connections
}
\description{
Establish connections to multiple database tables
}
