% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pandora_connection.R
\name{get_pandora_connection}
\alias{get_pandora_connection}
\title{Get the connection object for the Pandora DB}
\usage{
get_pandora_connection(cred_file = ".credentials")
}
\arguments{
\item{cred_file}{A credentials file containing three lines
listing host, user and password, respectively}
}
\value{
The Pandora DB connection object.
}
\description{
Get the connection object for the Pandora DB
}
