% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_table_formatting.R
\name{make_column_mandatory}
\alias{make_column_mandatory}
\title{make_column_mandatory}
\usage{
make_column_mandatory(x)
}
\arguments{
\item{x}{column name to wrap with asterisks}
}
\description{
For 'batch' uploading to update entries using the Pandora webpage, certain 
columns are mandatory. These are indicated in the header with asterisks. 
This function wraps a column name in the asterisks.
}
\details{
e.g. \*Sample_Id\* ,Archaeological ID,<...> where the first column must be 
supplied and the second not.
}
