% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_helper_functions.R
\name{entity2table}
\alias{entity2table}
\alias{table2entity}
\alias{convert_entity_table_name}
\title{name_conversion}
\usage{
entity2table(entity_type)

table2entity(table_name)

convert_entity_table_name(entity_type = c(), table_name = c())
}
\arguments{
\item{entity_type}{a valid entity type (e.g. site, sample, individual etc.)}

\item{table_name}{a valid table name (e.g. TAB_Site, TAB_Sample etc.)}
}
\description{
These functions converts a sidora entity type (e.g. site) to the 
corresponding PANDORA API table name (e.g. 'TAB_Site') and vice versa.
}
