% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_lookup_in_pandora.R
\name{namecol_value_from_id}
\alias{namecol_value_from_id}
\title{namecol_value_from_id}
\usage{
namecol_value_from_id(sidora_col_name, query_id, con, cache_dir = tempdir())
}
\arguments{
\item{sidora_col_name}{character. A sidora table column name}

\item{query_id}{integer vector. ID(s) to be converted to the human readable 'string' version}

\item{con}{a pandora connection}

\item{cache_dir}{a cache directory}
}
\description{
Given a sidora column name and a 'Id' integer will get the 
requested corresponding 'human readable' string version of the Id.
}
\details{
For example, given the ID 38 and the information that this ID was found in 
'extract.Batch', would result in Ex06_KE_2015-11-19
}
\examples{
\dontrun{
namecol_value_from_id(sidora_col_name = "extract.Batch", query_id = 38, con = con)
}

}
