% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidora.core.R
\docType{package}
\name{sidora.core-package}
\alias{sidora.core}
\alias{sidora.core-package}
\title{sidora.core: Basic Interface for PANDORA (MPI-SHH DAG)}
\description{
A basic interface to access data from the internal database PANDORA (MPI-SHH DAG).
}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID})

Authors:
\itemize{
  \item James Fellows Yates \email{jfy133@gmail.com} (\href{https://orcid.org/0000-0001-5585-6277}{ORCID})
}

}
\keyword{internal}
